clientside_v2/pyproject.toml
[tool.poetry]
name = "captureflow-agent"
version = "0.0.10"
description = "The CaptureFlow Tracer is a Python package crafted for in-depth tracing of function calls within Python applications. Its primary function is to capture and relay execution data to the CaptureFlow server-side system for decision making."
authors = ["Nick Kutz <me@nikitakuts.com>"]
readme = "README.md"
packages = [{include = "captureflow"}]

[tool.poetry.dependencies]
python = "^3.9"
opentelemetry-api = "^1.25.0"
opentelemetry-instrumentation = "^0.46b0"
opentelemetry-sdk = "^1.25.0"
python-dotenv = "1.0.1"
opentelemetry-exporter-otlp = "^1.25.0"
opentelemetry-instrumentation-fastapi = "^0.46b0"
opentelemetry-instrumentation-requests = "^0.46b0"
opentelemetry-instrumentation-httpx = "^0.46b0"
opentelemetry-instrumentation-flask = "^0.46b0"
opentelemetry-instrumentation-sqlalchemy = "^0.46b0"
opentelemetry-instrumentation-sqlite3 = "^0.46b0"
SQLAlchemy = "^2.0.30"
opentelemetry-instrumentation-dbapi = "^0.46b0"
wrapt = "^1.16.0"
opentelemetry-instrumentation-redis = "^0.46b0"

[tool.poetry.dev-dependencies]
fastapi = "^0.111.0"
httpx = "^0.27.0"
pytest = "^8.2.2"
redis = "^5.0.6"
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."opentelemetry_distro"]
distro = "captureflow.distro:CaptureFlowDistro"

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
